refactor $47 : eventlistener위치 변경
