+++ src/main/java/com/irum/orderservice/domain/order/event/CouponAppliedEvent.java
@@ -1,4 +1,4 @@
-package com.irum.orderservice.domain.order.event.event;
+package com.irum.orderservice.domain.order.event;
 
 import java.util.List;
 import java.util.UUID;
